requirements.txt
attrs==19.3.0
certifi==2020.4.5.1
more-itertools==8.2.0
numpy==1.18.4
packaging==20.3
Pillow==7.1.2
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.7
pytest==5.4.2
six==1.14.0
wcwidth==0.1.9
